refactor(hunyuan_image3): extract prompt_utils as shared builder
Move build_prompt and build_prompt_tokens out of the example script into
vllm_omni/diffusion/models/hunyuan_image3/prompt_utils.py so the AR-prefill
prompt template has a single source of truth that downstream callers can
reuse. The DiT pipeline keeps using TokenizerWrapper.apply_chat_template
(which eagerly consumes JointImageInfo); prompt_utils targets the lighter
client-side flow that uses an <img> placeholder + multi_modal_data.

README is updated to describe the actual instruct chat template (the
previous "pretrain template" wording was stale relative to the post-fix
behavior introduced earlier in this PR) and to point at the new module.

Addresses GH PR review comment requesting a common prompt-construction
function shared across AR / DiT / end2end.py.
